requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.3.0
azure-ai-inference==1.0.0b3
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0